Refactor : 자체 리팩토링
