+++ src/pages/kimnahee/Login/Login.scss
@@ -11,7 +11,7 @@ $primary-font: 'Lobster', cursive;
   align-items: center;
 }
 
-#loginBox {
+#login {
   width: 550px;
   border: $primary-border;
   margin: 10% auto;
