+++ src/pages/kimnahee/Main/Feed.js
@@ -1,18 +1,5 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import Comment from './CommentList';
-
-const FeedLists = () => {
-  const [feedList, setFeedList] = useState([]);
-
-  useEffect(() => {
-    fetch('/data/data.json')
-      .then(response => response.json())
-      .then(data => setFeedList(data));
-  }, []);
-
-  const Feeds = feedList.map(feed => <Feed key={feed.id} feed={feed} />);
-  return <div>{Feeds}</div>;
-};
 
 const Feed = ({ feed }) => {
   const [comment, setComment] = useState([]);
@@ -27,7 +14,7 @@ const Feed = ({ feed }) => {
   const addComment = e => {
     e.preventDefault();
     const commentElement = [
-      { id: keyId, userName: 'dlatldyd_dlatl', userContent: inputValue },
+      { id: keyId, userName: 'heehee.hee', userContent: inputValue },
     ];
     const newComment = comment.concat(commentElement);
     setKeyId(keyId + 1);
@@ -100,7 +87,7 @@ const Feed = ({ feed }) => {
           <div className="likeNumber">
             <img
               className="likeFeedUser"
-              alt="user image"
+              alt="userImage"
               src="https://plus.unsplash.com/premium_photo-1666788168130-959f109c07fc?ixlib=rb-4.0.3&amp;ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&amp;auto=format&amp;fit=crop&amp;w=822&amp;q=80"
             />
             <span>
@@ -129,4 +116,4 @@ const Feed = ({ feed }) => {
   );
 };
 
-export default FeedLists;
+export default Feed;
